Add code histogramming and plotting embedded mu+mu samples
These are samples in which the muons are replaced by muons instead of
taus. The code also supports replacement of generator-level muons in which
case for example the Zmumu selection efficiency corrections are not applied.

Included is also a python script to regenerate the Zmumu selection efficiency
correction root files.